tools/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250706
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
beautifulsoup4==4.13.4
blake3==1.0.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
colpali_engine==0.3.11.dev7+g481eb2008
compressed-tensors==0.9.2
cupy-cuda12x==13.5.1
dataclasses-json==0.6.7
datasets==4.0.0
decord==0.6.0
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.18.0
filetype==1.2.0
flash_attn==2.8.2
frozenlist==1.7.0
fsspec==2024.6.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lark==1.2.2
latex2sympy2_extended==1.10.2
liger_kernel==0.6.1
llama-cloud==0.1.5
llama-index==0.12.0
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.1
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.1
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.14
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
math-verify==0.8.0
mdurl==0.1.2
mistral_common==1.8.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.98.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-exporter-prometheus==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
peft==0.15.2
pillow==11.0.0
pip==25.1
platformdirs==4.3.8
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic-extra-types==2.10.5
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.2
pyzmq==27.0.1
qwen-vl-utils==0.0.8
ray==2.48.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.34.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
streamlit==1.47.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tensordict==0.9.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.33.0
vllm==0.8.2
wandb==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
